% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tutorials.R
\name{pull_tutorials}
\alias{pull_tutorials}
\title{Pull the USER 2020 Metaflow tutorials to the current folder}
\usage{
pull_tutorials()
}
\description{
Pull the USER 2020 Metaflow tutorials to the current folder
}
